#119: Allow entity objects to be fetched from plural field names
- This feature will be needed when there's camelCase field names for related entities, e.g. for user input on include filters
- This change also moves the location of get_entity_object_from_name() to prevent any issues with circular imports. Since the function no longer makes use of globals(), there's no requirement for the function to be in models.py
